New way of searching for resources.
We now use the fasit search api to search for queue names. This is more efficient then getting all ques matching one env class and then do an in memory filter.
Search api gives a short list of resources matching the search query. We also look up each resource and map it to the correct type. We also apply a filter to make sure we only get resources matching the type we want. In this case MQ Queue